Twitter/GachiKoi/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
future==0.18.2
idna==2.9
oauthlib==3.1.0
requests==2.23.0
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.11
